+++ packages/react-ui/src/widgets/comparative/ComparativeCategoryWidgetUI/ComparativeCategoryWidgetUI.js
@@ -60,6 +60,7 @@ function SearchIcon() {
  * @param {string[]} [props.selectedCategories]
  * @param {(categories: string[]) => any} [props.onSelectedCategoriesChange]
  * @param {(v: any) => any} [props.formatter]
+ * @param {boolean} [props.tooltip]
  * @param {(v: any) => any} [props.tooltipFormatter]
  * -->
  */
@@ -77,6 +78,7 @@ function ComparativeCategoryWidgetUI({
   selectedCategories = EMPTY_ARRAY,
   onSelectedCategoriesChange = IDENTITY_FN,
   formatter = IDENTITY_FN,
+  tooltip = true,
   tooltipFormatter = IDENTITY_FN
 }) {
   const theme = useTheme();
@@ -299,6 +301,7 @@ function ComparativeCategoryWidgetUI({
             filterable={filterable}
             formatter={formatter}
             tooltipFormatter={tooltipFormatter}
+            tooltip={tooltip}
             onClick={clickHandler}
             names={names}
           />
@@ -344,6 +347,7 @@ ComparativeCategoryWidgetUI.defaultProps = {
   searchable: true,
   filterable: true,
   selectedCategories: [],
+  tooltip: true,
   onSelectedCategoriesChange: IDENTITY_FN,
   formatter: IDENTITY_FN,
   tooltipFormatter: IDENTITY_FN
@@ -367,6 +371,7 @@ ComparativeCategoryWidgetUI.propTypes = {
   searchable: PropTypes.bool,
   filterable: PropTypes.bool,
   selectedCategories: PropTypes.arrayOf(PropTypes.string),
+  tooltip: PropTypes.bool,
   onSelectedCategoriesChange: PropTypes.func,
   formatter: PropTypes.func,
   tooltipFormatter: PropTypes.func